parallel socp solve on nyu, data logging
Using a parallel version of socp solve to speed up computation.
Because it's infeasible to solve a socp problem with 20M parameters,
we do a per-module computation of a new conflicting gradient.
Problem encountered: memory leak. Fixed by saving values of gradient to csv
and reading said csv in matlab through eng.eval.